full

+++ app/Repositories/Content/Content.php
@@ -5,37 +5,36 @@
 
 namespace App\Repositories\Content;
 
-use App\Repositories\Content\Drivers\File;
 use App\Repositories\Content\Exceptions\NotFoundException;
 use App\Repositories\Content\Exceptions\RouteExistsException;
 
+/**
+ * Class Content
+ * @package App\Repositories\Content
+ * @method bool save()
+ */
 class Content
 {
     const STATUS_DRAFT = 'draft';
     const STATUS_PUBLISHED = 'published';
 
     /**
-     * @var File
+     * @var DriverContract
      */
     protected $driver;
 
-    public function __construct($driver)
+    public function __construct(DriverContract $driver)
     {
         $this->driver = $driver;
     }
 
-    public function collection()
-    {
-        return $this->driver->collection();
-    }
-
-    public function create(array $data,$id=null)
+    public function create(array $data, $id = null)
     {
         if ($this->driver->exists(['route' => $data['route']])) {
             throw new RouteExistsException('Content already exists!');
         }
 
-        return $this->driver->create($data,$id);
+        return $this->driver->create($data, $id);
     }
 
     public function update($primary, array $data)
@@ -49,21 +48,11 @@ class Content
 
     public function delete($primary)
     {
+        if (!$this->driver->exists($primary)) {
+            throw new NotFoundException('Content not found!');
+        }
+
         return $this->driver->delete($primary);
-    }
-
-    /**
-     * @param $primary
-     * @return array
-     */
-    public function read($primary)
-    {
-        return $this->driver->read($primary);
-    }
-
-    public function exists($primary)
-    {
-        return $this->driver->exists($primary);
     }
 
     public function __call($method, $arguments)
